refactor: separate REST mapper classes
Separates V2 REST mapper concerns into dedicated classes:

- The general search query mappers remain in their place.
- Handling sorting moves to SearchQuerySortRequestMapper completely.
- Handling filters moves to a new SearchQueryFilterMapper class.
