+++ zeebe/gateway-rest/src/main/java/io/camunda/zeebe/gateway/rest/controller/authentication/AuthenticationController.java
@@ -7,7 +7,7 @@
  */
 package io.camunda.zeebe.gateway.rest.controller.authentication;
 
-import static io.camunda.zeebe.gateway.rest.SearchQueryResponseMapper.toCamundaUser;
+import static io.camunda.zeebe.gateway.rest.mapper.search.SearchQueryResponseMapper.toCamundaUser;
 
 import io.camunda.authentication.service.CamundaUserService;
 import io.camunda.spring.utils.ConditionalOnSecondaryStorageEnabled;
