+++ zeebe/gateway-rest/src/main/java/io/camunda/zeebe/gateway/rest/mapper/RestErrorMapper.java
@@ -5,7 +5,7 @@
  * Licensed under the Camunda License 1.0. You may not use this file
  * except in compliance with the Camunda License 1.0.
  */
-package io.camunda.zeebe.gateway.rest;
+package io.camunda.zeebe.gateway.rest.mapper;
 
 import io.camunda.service.exception.ServiceException;
 import io.camunda.service.exception.ServiceException.Status;
