+++ zeebe/gateway-rest/src/main/java/io/camunda/zeebe/gateway/rest/validator/UserTaskRequestValidator.java
@@ -17,7 +17,7 @@ import static io.camunda.zeebe.protocol.record.RejectionType.INVALID_ARGUMENT;
 import io.camunda.zeebe.gateway.protocol.rest.Changeset;
 import io.camunda.zeebe.gateway.protocol.rest.UserTaskAssignmentRequest;
 import io.camunda.zeebe.gateway.protocol.rest.UserTaskUpdateRequest;
-import io.camunda.zeebe.gateway.rest.RestErrorMapper;
+import io.camunda.zeebe.gateway.rest.mapper.RestErrorMapper;
 import java.util.List;
 import java.util.Optional;
 import org.springframework.http.HttpStatus;
